moved the flattenTOC function to utils
