+++ src/data/constitution.tsx
@@ -1,4 +1,5 @@
 import * as Constitution from '../assets/data/constitution.json';
+import { flattenTOC } from "../utils";
 
 export const constitutionData: any = (Constitution as any).default;
 // wrap the entire content in a div so the body has a single child
@@ -33,19 +34,6 @@ export function getTOCEntry(id: string) {
 
     search(tableOfContents);
     return match;
-}
-
-function flattenTOC (arr: any[]) {
-    let result: string[] = [];
-
-    arr.forEach(el => {
-        if (el.id) result.push(el.id);
-        if (el.children && el.children.length > 0) {
-            result = result.concat(flattenTOC(el.children));
-        };
-    });
-
-    return result;
 }
 
 export const flatTOC = flattenTOC(tableOfContents);
